Moved the camel_to_snake function to the .eds file, attempted to add a comment to the parse function in canopen.py
